Implement connect user and video

+++ src/controllers/videoController.js
@@ -1,5 +1,5 @@
 import videoModel from "../models/video";
-
+import userModel from "../models/user";
 /* //example of using callback method
   videoModel.find({}, (error, videos) => {
     if (error) {
@@ -19,28 +19,48 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await videoModel.findById(id);
+  const video = await videoModel.findById(id).populate("owner");
   if (!video)
     return res.status(404).render("404", { pageTitle: "Video Not Found." });
 
-  return res.render("watch", { pageTitle: video.title, video });
+  return res.render("watch", {
+    pageTitle: video.title,
+    video,
+  });
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
   const video = await videoModel.findById(id);
   if (!video)
     return res.status(404).render("404", { pageTitle: "Video Not Found." });
 
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `Editing ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+
   const { title, description, hashtags } = req.body;
-  const video = await videoModel.exists({ _id: id });
+
+  const video = await videoModel.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
 
   await videoModel.findByIdAndUpdate(id, {
@@ -48,8 +68,6 @@ export const postEdit = async (req, res) => {
     description,
     hashtags: videoModel.formatHashtags(hashtags),
   });
-  // hashtags: hashtags.split(",").map((word) => (word.startsWith("#") ? word : `#${word}`))
-
   return res.redirect(`/videos/${id}`);
 };
 
@@ -59,16 +77,24 @@ export const getUpload = (req, res) => {
 
 export const postUpload = async (req, res) => {
   const {
+    session: {
+      user: { _id },
+    },
     body: { title, description, hashtags },
     file: { path: fileUrl },
   } = req;
   try {
-    await videoModel.create({
+    const newVideo = await videoModel.create({
       title,
       description,
       fileUrl,
+      owner: _id,
       hashtags: videoModel.formatHashtags(hashtags),
     });
+    const user = await userModel.findById(_id);
+    user.videos.unshift(newVideo._id);
+    user.save();
+
     return res.redirect("/");
   } catch (error) {
     return res.render("upload", {
@@ -80,7 +106,24 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+
+  const video = await videoModel.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
   await videoModel.findByIdAndDelete(id);
+  const user = await userModel.findById(_id);
+
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
 
@@ -93,7 +136,6 @@ export const search = async (req, res) => {
         title: { $regex: new RegExp(keyword, "i") },
       })
       .sort({ createdAt: "desc" });
-    console.log(videos);
   }
   return res.render("search", { pageTitle: "Search", videos });
 };
